--- src/core-extensions/meta-types.ts
@@ -1,4 +1,0 @@
-export interface GluegunMeta {
-  version: () => string
-  commandInfo: () => string[][]
-}
