+++ src/core-extensions/patching-extension.ts
@@ -1,5 +1,5 @@
 import { GluegunToolbox } from '../domain/toolbox'
-import { update, append, prepend, replace, patch, exists } from '../toolbox/patching-tools'
+import { patching } from '../toolbox/patching-tools'
 
 /**
  * Builds the patching feature.
@@ -7,5 +7,5 @@ import { update, append, prepend, replace, patch, exists } from '../toolbox/patc
  * @param toolbox The running toolbox.
  */
 export default function attach(toolbox: GluegunToolbox): void {
-  toolbox.patching = { update, append, prepend, replace, patch, exists }
+  toolbox.patching = patching
 }
