+++ src/loaders/module-loader.ts
@@ -1,12 +1,12 @@
-import { isNotFile } from '../toolbox/filesystem-tools'
-import { isBlank } from '../toolbox/string-tools'
+import { filesystem } from '../toolbox/filesystem-tools'
+import { strings } from '../toolbox/string-tools'
 
 // try loading this module
 export function loadModule(path) {
-  if (isBlank(path)) {
+  if (strings.isBlank(path)) {
     throw new Error('path is required')
   }
-  if (isNotFile(path)) {
+  if (filesystem.isNotFile(path)) {
     throw new Error(`${path} is not a file`)
   }
