+++ src/toolbox/http-tools.ts
@@ -1,0 +1,6 @@
+import { create } from 'apisauce'
+import { GluegunHttp } from './http-types'
+
+const http: GluegunHttp = { create }
+
+export { http, GluegunHttp }
