+++ src/toolbox/semver-tools.ts
@@ -1,0 +1,6 @@
+import * as semverTool from 'semver'
+import { GluegunSemver } from './semver-types'
+
+const semver: GluegunSemver = semverTool
+
+export { semver, GluegunSemver }
