+++ src/toolbox/strings-types.ts
